+++ plugins/db-wakamiti-plugin/src/main/java/es/iti/wakamiti/database/jdbc/Database.java
@@ -8,6 +8,8 @@ package es.iti.wakamiti.database.jdbc;
 
 import com.mdimension.jchronic.Chronic;
 import com.mdimension.jchronic.Options;
+import es.iti.wakamiti.api.WakamitiException;
+import es.iti.wakamiti.api.WakamitiStepRunContext;
 import es.iti.wakamiti.api.util.WakamitiLogger;
 import es.iti.wakamiti.database.SQLParser;
 import es.iti.wakamiti.database.exception.SQLRuntimeException;
@@ -16,14 +18,16 @@ import org.slf4j.Logger;
 import java.math.BigDecimal;
 import java.math.BigInteger;
 import java.sql.*;
+import java.time.Instant;
+import java.time.LocalDate;
 import java.time.LocalDateTime;
+import java.time.format.DateTimeParseException;
 import java.time.temporal.ChronoUnit;
 import java.util.*;
 import java.util.function.UnaryOperator;
 import java.util.stream.Stream;
 
-import static es.iti.wakamiti.database.DatabaseHelper.DATE_TIME_FORMATTER;
-import static es.iti.wakamiti.database.DatabaseHelper.collectToMap;
+import static es.iti.wakamiti.database.DatabaseHelper.*;
 import static es.iti.wakamiti.database.jdbc.LogUtils.*;
 import static java.util.Objects.isNull;
 
@@ -298,12 +302,34 @@ public final class Database {
             case TIME:
             case TIME_WITH_TIMEZONE:
             case TIMESTAMP_WITH_TIMEZONE:
-                Calendar calendar2 = Chronic.parse(value, new Options(false)).getBeginCalendar();
-                value = DATE_TIME_FORMATTER.withZone(calendar2.getTimeZone().toZoneId()).format(calendar2.toInstant());
+                try {
+                    LocalDateTime dateTime = parse(value, LocalDateTime.class);
+                    value = DATE_TIME_FORMATTER.format(dateTime);
+                } catch (WakamitiException | DateTimeParseException e) {
+                    Calendar calendar2 = Chronic.parse(value, new Options(false)).getBeginCalendar();
+                    value = DATE_TIME_FORMATTER.withZone(calendar2.getTimeZone().toZoneId()).format(calendar2.toInstant());
+                }
                 return WakamitiTimestamp.valueOf(value, false);
             default:
                 return value;
         }
+    }
+
+    /**
+     * Parses a text expression into the specified data type.
+     *
+     * @param expression The text expression to parse.
+     * @param type       The class of the type into which the expression should be parsed.
+     * @param <T>        The type parameter representing the target type.
+     * @return The parsed value of the specified type.
+     * @throws WakamitiException if no type registry is found for the specified class.
+     */
+    @SuppressWarnings("unchecked")
+    private <T> T parse(String expression, Class<T> type) {
+        WakamitiStepRunContext ctx = WakamitiStepRunContext.current();
+        return (T) ctx.typeRegistry().findTypesForJavaType(type).findFirst()
+                .orElseThrow(() -> new WakamitiException("No type registry found for Class '{}'", type))
+                .parse(ctx.stepLocale(), expression);
     }
 
     /**