+++ plugins/db-wakamiti-plugin/src/main/java/es/iti/wakamiti/database/jdbc/DatabaseType.java
@@ -16,36 +16,38 @@ public enum DatabaseType {
 
 
 
-    ORACLE("select 1 from dual"),
-    HSQLDB("SELECT 1 FROM INFORMATION_SCHEMA.SYSTEM_USERS"),
-    H2(),
-    SQLSERVER(),
-    MYSQL(),
-    MARIADB(),
-    POSTGRESQL(),
-    SQLITE(),
-    DB2("select 1 from sysibm.sysdummy1"),
-    AS400("select 1 from sysibm.sysdummy1"),
-    DERBY("SELECT 1 FROM SYSIBM.SYSDUMMY1"),
-    INFORMIX("select count(*) from systables"),
-    OTHER();
-
-    private static final String DEFAULT = "select 1";
+    ORACLE("select 1 from dual", "select DBTIMEZONE from dual"),
+    HSQLDB("SELECT 1 FROM INFORMATION_SCHEMA.SYSTEM_USERS",
+            "SELECT TIMEZONE() FROM INFORMATION_SCHEMA.SYSTEM_USERS"),
+    H2("select 1", "SELECT @@system_time_zone"),
+    SQLSERVER("select 1", "SELECT CONVERT(DATETIME, LEFT(SYSDATETIMEOFFSET(),19), 126)"),
+    MYSQL("select 1", "SELECT @@system_time_zone"),
+    MARIADB("select 1", "SELECT @@system_time_zone"),
+    POSTGRESQL("select 1", "select to_char(CURRENT_TIMESTAMP(), 'TZ')"),
+    SQLITE("select 1", ""),
+    DB2("select 1 from sysibm.sysdummy1", ""),
+    AS400("select 1 from sysibm.sysdummy1", ""),
+    DERBY("SELECT 1 FROM SYSIBM.SYSDUMMY1", ""),
+    INFORMIX("select count(*) from systables", ""),
+    OTHER("select 1", "");
 
     private final String healthCheckSql;
+    private final String timezoneSql;
 
     /**
      * Constructs a DatabaseType enum constant with the given health check SQL query.
      *
      * @param healthCheckSql the SQL query used for health check
      */
-    DatabaseType(String healthCheckSql) {
+    DatabaseType(String healthCheckSql, String timezoneSql) {
         this.healthCheckSql = healthCheckSql;
+        this.timezoneSql = timezoneSql;
     }
 
-    DatabaseType() {
-        this.healthCheckSql = DEFAULT;
-    }
+//    DatabaseType() {
+//        this.healthCheckSql = "select 1";
+//        this.timezoneSql = "SELECT @@system_time_zone;";
+//    }
 
     /**
      * Retrieves the DatabaseType enum constant based on the provided JDBC URL.